Rebase with master and fix the schema migration
